move ocr logic into services
